# crates/analytics-api/openapi/aux-analytics.openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "description": "Error response for invalid requests and query failures.",
        "example": {
          "error": "query must be a single read-only SELECT statement"
        },
        "properties": {
          "error": {
            "description": "Human-readable error message.",
            "example": "query must be a single read-only SELECT statement",
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "IngestBatchResponse": {
        "description": "Batch stream ingestion outcome.",
        "example": {
          "outcomes": [
            {
              "outcome": "inserted"
            },
            {
              "outcome": "updated"
            }
          ],
          "record_count": 2
        },
        "properties": {
          "outcomes": {
            "description": "Per-record ingest outcomes in request order.",
            "items": {
              "$ref": "#/components/schemas/IngestResponse"
            },
            "type": "array"
          },
          "record_count": {
            "description": "Number of records accepted by the engine transaction.",
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "record_count",
          "outcomes"
        ],
        "type": "object"
      },
      "IngestResponse": {
        "description": "Stream ingestion outcome.",
        "example": {
          "outcome": "inserted"
        },
        "properties": {
          "outcome": {
            "description": "Final ingestion action performed by the engine.",
            "example": "inserted",
            "type": "string"
          },
          "privacy_dropped_fields": {
            "default": null,
            "description": "Count of fields or scalar values removed by the privacy policy.",
            "example": 2,
            "format": "int64",
            "minimum": 0,
            "type": [
              "integer",
              "null"
            ]
          },
          "privacy_policy_version": {
            "default": null,
            "description": "Active privacy policy version, when privacy mode filtered this request.",
            "example": "privacy-v1",
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "outcome"
        ],
        "type": "object"
      },
      "IngestStreamRecordBatchRequest": {
        "description": "Batch stream ingestion request.",
        "example": {
          "records": [
            {
              "record": {
                "Keys": {
                  "pk": {
                    "S": "USER#user-1"
                  }
                },
                "NewImage": {
                  "org_id": {
                    "S": "org-a"
                  },
                  "pk": {
                    "S": "USER#user-1"
                  },
                  "profile": {
                    "M": {
                      "email": {
                        "S": "ada@example.com"
                      }
                    }
                  }
                },
                "SequenceNumber": "49668899999999999999999999999999999999999999999999999999"
              },
              "record_key": "user-1"
            }
          ]
        },
        "properties": {
          "records": {
            "description": "Stream records from one source poll response. Keep payloads within the\nsource service response limit, such as `DynamoDB` Streams' 1 MiB page\ncap.",
            "items": {
              "$ref": "#/components/schemas/IngestStreamRecordRequest"
            },
            "minItems": 1,
            "type": "array"
          }
        },
        "required": [
          "records"
        ],
        "type": "object"
      },
...
      "IngestStreamRecordRequest": {
        "description": "Stream ingestion request.",
        "example": {
          "record": {
            "Keys": {
              "pk": {
                "S": "USER#user-1"
              }
            },
            "NewImage": {
              "org_id": {
                "S": "org-a"
              },
              "pk": {
                "S": "USER#user-1"
              },
              "profile": {
                "M": {
                  "email": {
                    "S": "ada@example.com"
                  }
                }
              }
            },
            "SequenceNumber": "49668899999999999999999999999999999999999999999999999999"
          },
          "record_key": "user-1"
        },
        "properties": {
          "record": {
            "$ref": "#/components/schemas/IngestStreamRecordPayload",
            "description": "Stream record in the analytics wire format or a standard stream event\nenvelope."
          },
          "record_key": {
            "default": null,
            "description": "Stable row key for idempotent writes. Defaults to the stream sequence or\nevent id when omitted.",
            "example": "user-1",
            "maxLength": 1024,
            "minLength": 1,
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "record"
        ],
        "type": "object"
      },
...
      "OperationListResponse": {
        "properties": {
          "operations": {
            "items": {
              "$ref": "#/components/schemas/OperationStatusResponse"
            },
            "type": "array"
          }
        },
        "required": [
          "operations"
        ],
        "type": "object"
      },
...
      "TenantQueryRequest": {
        "description": "Tenant-scoped structured query request for callers that should not\nconstruct SQL.",
        "example": {
          "query": {
            "analytics_table_name": "users",
            "filters": [
              {
                "expression": {
                  "column_name": "org_id",
                  "kind": "column"
                },
                "kind": "eq",
                "value": "org-a"
              }
            ],
            "group_by": [],
            "limit": 100,
            "order_by": [
              {
                "direction": "asc",
                "expression": {
                  "column_name": "email",
                  "kind": "column"
                }
              }
            ],
            "select": [
              {
                "alias": null,
                "column_name": "email",
                "kind": "column"
              }
            ]
          },
          "target_tenant_id": "tenant_01"
        },
        "properties": {
          "query": {
            "$ref": "#/components/schemas/StructuredQuery",
            "description": "Structured query tree compiled by the analytics engine."
          },
          "target_tenant_id": {
            "description": "Tenant id that the query is authorized to read. The API injects this as\nan engine-owned tenant predicate; callers must not rely on their own\ntenant filters for isolation.",
            "example": "tenant_01",
            "maxLength": 255,
            "minLength": 1,
            "type": "string"
          }
        },
        "required": [
          "target_tenant_id",
          "query"
        ],
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "",
    "license": {
      "identifier": "BUSL-1.1",
      "name": "BUSL-1.1"
    },
    "title": "analytics-api",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/ingest-batch/{analytics_table_name}": {
      "post": {
        "operationId": "ingest_stream_record_batch",
        "parameters": [
          {
            "description": "Registered analytics table name",
            "in": "path",
            "name": "analytics_table_name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestStreamRecordBatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestBatchResponse"
                }
              }
            },
            "description": "Batch ingest outcomes"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request or batch ingest failure"
          }
        },
        "tags": [
          "Analytics"
        ]
      }
    },
    "/ingest/{analytics_table_name}": {
      "post": {
        "operationId": "ingest_stream_record",
        "parameters": [
          {
            "description": "Registered analytics table name",
            "in": "path",
            "name": "analytics_table_name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/IngestStreamRecordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestResponse"
                }
              }
            },
            "description": "Ingest outcome"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request or ingest failure"
          }
        },
        "tags": [
          "Analytics"
        ]
      }
    },
...
    "/operations": {
      "get": {
        "operationId": "list_operations",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OperationListResponse"
                }
              }
            },
            "description": "Durable operation list"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Operation store is not configured"
          }
        },
        "tags": [
          "Operations"
        ]
      }
    },
...
    "/tenant-query": {
      "post": {
        "operationId": "tenant_query",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TenantQueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResponse"
                }
              }
            },
            "description": "Query tenant-scoped rows as JSON objects"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request or query"
          }
        },
        "tags": [
          "Analytics"
        ]
      }
    },
    "/unscoped-sql-query": {
      "post": {
        "operationId": "unscoped_sql_query",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnscopedSqlQueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Raw SQL is not available through HTTP without an admin boundary"
          }
        },
        "tags": [
          "Analytics"
        ]
      }
    },
    "/unscoped-structured-query": {
      "post": {
        "operationId": "unscoped_structured_query",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UnscopedStructuredQueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueryResponse"
                }
              }
            },
            "description": "Query explicitly unscoped rows as JSON objects"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Invalid request or query"
          }
        },
        "tags": [
          "Analytics"
        ]
      }
    },
...
  },
...
}